+++ src/main/java/com/togedy/togedy_server_v2/domain/study/scheduler/StudyTierScheduler.java
@@ -1,4 +1,4 @@
-package com.togedy.togedy_server_v2.domain.study.application.scheduler;
+package com.togedy.togedy_server_v2.domain.study.scheduler;
 
 import com.togedy.togedy_server_v2.domain.study.application.StudyTierService;
 import lombok.RequiredArgsConstructor;
@@ -7,13 +7,12 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class StudyStatisticsScheduler {
+public class StudyTierScheduler {
 
     private final StudyTierService studyTierService;
 
-    @Scheduled(cron = "0 30 5 * * *", zone = "Asia/Seoul")
-    public void calculateChallengeStudyScores() {
-        studyTierService.calculateChallengeStudyScores();
+    @Scheduled(cron = "0 0 6 * * *", zone = "Asia/Seoul")
+    public void updateStudyTier() {
+        studyTierService.applyStudyTier();
     }
-
 }
